+++ server/modules/monitor/mariadbmon/mariadbmon.hh
@@ -487,7 +487,8 @@ private:
         std::chrono::seconds ssh_timeout;           /**< SSH connection and command timeout */
         int64_t              ssh_port {0};          /**< SSH port on all servers */
         int64_t              rebuild_port {0};      /**< Listen port for server backup transfer */
-        std::string          mbu_use_memory;        /**< Given to mariabackup --prepare */
+        std::string          mbu_use_memory;        /**< Given to mariabackup --prepare --use-memory=<val> */
+        int64_t              mbu_parallel {1};      /**< Given to mariabackup --parallel=<val> */
         std::string          backup_storage_addr;   /**< Backup storage host */
         std::string          backup_storage_path;   /**< Backup storage directory */